[rui-radio] rename file
