+++ packages/perseus-editor/src/widgets/__tests__/radio-editor.test.tsx
@@ -5,7 +5,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
 import {testDependencies} from "../../../../../testing/test-dependencies";
-import RadioEditor from "../radio/editor";
+import RadioEditor from "../radio/radio-editor";
 
 import type {PerseusRadioChoice} from "@khanacademy/perseus-core";
 import type {UserEvent} from "@testing-library/user-event";
